// src/lib/scoretrakapis/scoretrak/host_group/v1/host_group_pb.d.ts
  export type AsObject = {
    id?: scoretrak_proto_v1_uuid_pb.UUID.AsObject;
    name: string;
    hide?: google_protobuf_wrappers_pb.BoolValue.AsObject;
    pause?: google_protobuf_wrappers_pb.BoolValue.AsObject;
    hosts?: scoretrak_host_v1_host_pb.Host.AsObject;
  };
// ...
export class GetAllRequest extends jspb.Message {
  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): GetAllRequest.AsObject;
  static toObject(
    includeInstance: boolean,
    msg: GetAllRequest
  ): GetAllRequest.AsObject;
  static serializeBinaryToWriter(
    message: GetAllRequest,
    writer: jspb.BinaryWriter
  ): void;
  static deserializeBinary(bytes: Uint8Array): GetAllRequest;
  static deserializeBinaryFromReader(
    message: GetAllRequest,
    reader: jspb.BinaryReader
  ): GetAllRequest;
}

export namespace GetAllRequest {
  export type AsObject = {};
}